requirements.txt
requests==2.31.0
beautifulsoup4==4.12.2
numpy==1.26.2
pandas==2.1.3
postgrest==0.13.0
supabase==2.3.0
urllib3==2.1.0
python-dotenv==1.0.0
